+++ src/mapper_erlang.ts
@@ -1,5 +1,10 @@
 "use strict";
 /**
+ * ------------------------------------------ 
+ * This file is a part of CodeMap distribution. 
+ * It will be overwritten after the extension next update. Thus you may want to make an editable copy of
+ * this file and add it as a custom dedicated mapper in the settings file.
+ * ------------------------------------------ 
  * @name            mapper_erlang
  * @author          Andrea Benini (andrea benini  at  google public mail)
  * @description     Simple and barebone erlang mapper for Visual Studio Code CodeMap plugin,
@@ -38,7 +43,7 @@ class mapper {
                 .read_all_lines(file)
                 .forEach(line => {
                     // First line, file name
-                    if (line_num==0) {
+                    if (line_num == 0) {
                         let basename = file.replace(/\\/g, '/').replace(/.*\//, '').replace(/(.+)(.erl)/, '$1');
                         members.push(`${basename}|1|class`);
                     }
@@ -46,26 +51,26 @@ class mapper {
                     line = line.trim();
                     // Include              '-include("SOMETHING")...'
                     let Include = line.match(/-include\s*\("\s*(.+)"(.+)/);
-                    if (Include!=null && Include[1]!=null && Include[1]!='') {
+                    if (Include != null && Include[1] != null && Include[1] != '') {
                         members.push(`    ${Include[1]}|${line_num}|document`);
                     }
                     // Behaviour            '-behaviour(SOMETHING)...'
                     let Behaviour = line.match(/-behaviour\s*\(\s*(.+)\s*\)(.+)/);
-                    if (Behaviour!=null && Behaviour[1]!=null && Behaviour[1]!='') {
+                    if (Behaviour != null && Behaviour[1] != null && Behaviour[1] != '') {
                         members.push(`    ${Behaviour[1]}|${line_num}|interface`);
                     }
                     // Define               "-define(SOMETHING, ..."
                     let Define = line.match(/^-define\s*\(\s*([\w0-9]+)(.+)/);
-                    if (Define!=null && Define[1]!=null && Define[1]!='') {
+                    if (Define != null && Define[1] != null && Define[1] != '') {
                         members.push(`    ${Define[1]}|${line_num}|property`);
                     }
                     // Function               [a-z][a-zA-Z0-9_](.+) ->
                     let Function = line.match(/^\s*([a-z][a-zA-Z0-9_]+)\s*(\(.*\))\s*(->)\s*(.*)$/);
-                    if (Function!=null && Function[1]!=null && Function[1]!='' && Function[3] == '->') {
+                    if (Function != null && Function[1] != null && Function[1] != '' && Function[3] == '->') {
                         Function[2] = Function[2].replace(/\|/g, String.fromCharCode(0x01C0));  // Dirty hack but "|" is reserved
                         members.push(`${Function[1]}${Function[2]}|${line_num}|function`);
                     }
-            });
+                });
         } catch (error) {
         }
         return members;